# Improve ExampeLines example