Each Line has a filenameIndex instead of containing the filename related fields.

Makes the overall PDB lines structure clearer and makes "Storing lines from modules " 1.3x faster

Before
====

Running example "Lines"
| Reading image section stream
| ---> done in 0.049ms
| Reading module info stream
| ---> done in 0.208ms
| Reading names stream
| ---> done in 6.324ms
| Storing lines from modules
| ---> done in 227.087ms (1815 elements)
| std::sort sections
| ---> done in 87.234ms (3895467 elements)

After 
===

Running example "Lines"
| Reading image section stream
| ---> done in 0.043ms
| Reading module info stream
| ---> done in 0.172ms
| Reading names stream
| ---> done in 5.428ms
| Storing lines from modules
| ---> done in 167.206ms (1815 elements)
| std::sort sections
| ---> done in 80.895ms (3895467 elements)